+++ fabric/src/main/java/com/lx862/jcm/mod/registry/Networking.java
@@ -20,16 +20,18 @@ public class Networking {
 
         // Block Entity Update
         registerPacket(ButterflyLightUpdatePacket.class, ButterflyLightUpdatePacket::new);
+        registerPacket(EnquiryUpdateGUIPacket.class, EnquiryUpdateGUIPacket::new);
         registerPacket(FareSaverUpdatePacket.class, FareSaverUpdatePacket::new);
+        registerPacket(OperatorButtonUpdatePacket.class, OperatorButtonUpdatePacket::new);
         registerPacket(PIDSUpdatePacket.class, PIDSUpdatePacket::new);
         registerPacket(PIDSProjectorUpdatePacket.class, PIDSProjectorUpdatePacket::new);
         registerPacket(SoundLooperUpdatePacket.class, SoundLooperUpdatePacket::new);
         registerPacket(SubsidyMachineUpdatePacket.class, SubsidyMachineUpdatePacket::new);
-        registerPacket(EnquiryUpdateGUIPacket.class, EnquiryUpdateGUIPacket::new);
 
         // GUI Screen
         registerPacket(ButterflyLightGUIPacket.class, ButterflyLightGUIPacket::new);
         registerPacket(FareSaverGUIPacket.class, FareSaverGUIPacket::new);
+        registerPacket(OperatorButtonGUIPacket.class, OperatorButtonGUIPacket::new);
         registerPacket(PIDSGUIPacket.class, PIDSGUIPacket::new);
         registerPacket(PIDSProjectorGUIPacket.class, PIDSProjectorGUIPacket::new);
         registerPacket(SoundLooperGUIPacket.class, SoundLooperGUIPacket::new);