+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## JCM (v2.0.0-beta.14) for (MTR 4.0.0-prerelease.1) has been released!
 
+**Changes:**
+- To adapt to the new train driver key changes in MTR 4, **Operator Button** now allows choosing which driver keys are allowed.
+- - Configurable via brush
+
 **Fix:**
 - Fix JSON PIDS ETA text potentially not following the destination's language
-
-**Mod Integration**
-- Add English support for [Item Descriptions](https://modrinth.com/mod/item-descriptions), short descriptions for JCM blocks will now be shown when using the mod!
 
 **Download:**  
 You can download this release on [Modrinth](https://modrinth.com/mod/jcm) or [GitHub](https://github.com/DistrictOfJoban/Joban-Client-Mod/releases)